+++ src/middleware.ts
@@ -3,7 +3,7 @@ import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 
 // Set the paths that don't require the user to be signed in
-const publicPaths = ["/sign-in*", "/sign-up*", "/about"];
+const publicPaths = ["/", "/sign-in*", "/sign-up*", "/about"];
 
 const isPublic = (path: string) => {
   return publicPaths.find((x) =>
